+++ src-tauri/service/src/models/invoice.rs
@@ -3,18 +3,21 @@ use super::invoice_item::{SelectInvoicesItems, SelectInvoicesItemsForUpdate};
 use sea_orm::FromQueryResult;
 use specta::Type;
 use serde::{Deserialize, Serialize};
+use sea_orm::entity::prelude::Decimal;
 
 #[derive(Deserialize, Serialize, Debug, PartialEq, FromQueryResult, Type)]
 pub struct SelectInvoices {
     pub id: String,
     pub created_at: String,
-    pub paid_amount: f64,
+    #[specta(type = f64)]
+    pub paid_amount: Decimal,
     pub client_id: String,
     pub full_name: String,
     pub status: String,
     pub identifier: String,
     pub products: i64,
-    pub total: f64,
+    #[specta(type = f64)]
+    pub total: Decimal,
 }
 
 #[derive(Deserialize, Serialize, Debug, PartialEq, FromQueryResult, Type)]
@@ -22,14 +25,16 @@ pub struct SelectInvoiceDetails {
     pub id: String,
     pub order_id: String,
     pub created_at: String,
-    pub paid_amount: f64,
+    #[specta(type = f64)]
+    pub paid_amount: Decimal,
     pub full_name: String,
     pub address: Option<String>,
     pub phone_number: Option<String>,
     pub email: Option<String>,
     pub status: String,
     pub identifier: String,
-    pub total: f64,
+    #[specta(type = f64)]
+    pub total: Decimal,
 }
 
 #[derive(Debug, Serialize, Deserialize, Type)]
@@ -37,7 +42,8 @@ pub struct NewInvoice {
     pub client_id: String,
     pub order_id: Option<String>,
     pub status: String,
-    pub paid_amount: f64,
+    #[specta(type = f64)]
+    pub paid_amount: Decimal,
     pub items: Vec<NewInvoiceItem>,
 }
 
@@ -46,7 +52,8 @@ pub struct UpdateInvoice {
     pub id: String,
     pub client_id: String,
     pub status: String,
-    pub paid_amount: f64,
+    #[specta(type = f64)]
+    pub paid_amount: Decimal,
     pub items: Vec<UpdateInvoiceItem>,
 }
 
@@ -59,8 +66,10 @@ pub struct InvoicesResponse {
 #[derive(Debug, Serialize, Deserialize, FromQueryResult, Type)]
 pub struct InvoiceProductItem {
     pub name: String,
-    pub price: f64,
-    pub quantity: f64,
+    #[specta(type = f64)]
+    pub price: Decimal,
+    #[specta(type = f64)]
+    pub quantity: Decimal,
 }
 
 #[derive(Debug, Serialize, Deserialize, Type)]
@@ -70,7 +79,8 @@ pub struct InvoiceWithClient {
     pub created_at: String,
     pub status: String,
     pub identifier: Option<String>,
-    pub paid_amount: f64,
+    #[specta(type = f64)]
+    pub paid_amount: Decimal,
     pub full_name: String,
     pub email: Option<String>,
     pub address: Option<String>,
@@ -84,8 +94,10 @@ pub struct InvoiceDetailsResponse {
     pub created_at: String,
     pub status: String,
     pub identifier: String,
-    pub paid_amount: f64,
-    pub total: f64,
+    #[specta(type = f64)]
+    pub paid_amount: Decimal,
+    #[specta(type = f64)]
+    pub total: Decimal,
     pub client: InvoiceClientInfo,
     pub items: Vec<SelectInvoicesItems>,
 }
