+++ src-tauri/service/src/models/product.rs
@@ -1,6 +1,7 @@
 use sea_orm::FromQueryResult;
 use specta::Type;
 use serde::{Deserialize, Serialize};
+use sea_orm::entity::prelude::Decimal;
 
 #[derive(Deserialize, Serialize, Debug, PartialEq, FromQueryResult, Type)]
 pub struct SelectProducts {
@@ -9,10 +10,14 @@ pub struct SelectProducts {
     pub created_at: String,
     pub description: Option<String>,
     pub image: Option<String>,
-    pub selling_price: Option<f64>,
-    pub purchase_price: Option<f64>,
-    pub inventory: f64,
-    pub min_quantity: Option<f64>,
+    #[specta(type = f64)]
+    pub selling_price: Option<Decimal>,
+    #[specta(type = f64)]
+    pub purchase_price: Option<Decimal>,
+    #[specta(type = f64)]
+    pub inventory: Decimal,
+    #[specta(type = f64)]
+    pub min_quantity: Option<Decimal>,
 }
 
 #[derive(Deserialize, Serialize, Debug, Type)]
@@ -25,16 +30,20 @@ pub struct ProductsResponse {
 pub struct ProductSearch {
     pub label: String,
     pub value: String,
-    pub price: Option<f64>,
+    #[specta(type = f64)]
+    pub price: Option<Decimal>,
 }
 
 #[derive(Debug, Serialize, Deserialize, Type)]
 pub struct NewProduct {
     pub name: String,
     pub description: Option<String>,
-    pub selling_price: f64,
-    pub purchase_price: f64,
-    pub min_quantity: f64,
+    #[specta(type = f64)]
+    pub selling_price: Decimal,
+    #[specta(type = f64)]
+    pub purchase_price: Decimal,
+    #[specta(type = f64)]
+    pub min_quantity: Decimal,
     pub image: Option<String>,
 }
 
@@ -43,9 +52,12 @@ pub struct Product {
     pub id: String,
     pub name: String,
     pub description: Option<String>,
-    pub selling_price: f64,
-    pub purchase_price: f64,
-    pub min_quantity: f64,
+    #[specta(type = f64)]
+    pub selling_price: Decimal,
+    #[specta(type = f64)]
+    pub purchase_price: Decimal,
+    #[specta(type = f64)]
+    pub min_quantity: Decimal,
     pub image: Option<String>,
 }
 
@@ -54,8 +66,11 @@ pub struct UpdateProduct {
     pub id: String,
     pub name: Option<String>,
     pub description: Option<String>,
-    pub selling_price: Option<f64>,
-    pub purchase_price: Option<f64>,
-    pub min_quantity: Option<f64>,
+    #[specta(type = f64)]
+    pub selling_price: Option<Decimal>,
+    #[specta(type = f64)]
+    pub purchase_price: Option<Decimal>,
+    #[specta(type = f64)]
+    pub min_quantity: Option<Decimal>,
     pub image: Option<String>,
 }
